refactored mpitests
