+++ tests/.run_tests.sh
@@ -4,8 +4,14 @@ set -e
 
 MPI_EXEC=$(python -c "import mpi4py;import os;filename = next(iter(mpi4py.get_config().items()))[1];print(os.path.dirname(filename)+'/mpiexec');")
 
-$MPI_EXEC -np 3 coverage run --parallel-mode --source=bingo tests/integration/mpitest_parallel_archipelago.py
-$MPI_EXEC -np 3 coverage run --parallel-mode --source=bingo tests/integration/mpitest_parallel_archipelago_and_multiprocessing.py
+for i in tests/integration/mpitests/*.py
+do
+  if [ $i != "tests/integration/mpitests/mpitest_util.py" ]
+  then
+    echo "Running mpitest: $i"
+    $MPI_EXEC -np 3 coverage run --parallel-mode --source=bingo $i
+  fi
+done
 coverage combine
 
 pytest tests --cov=bingo --cov-report=term-missing --cov-append
